Added a healthbar, powerups and an obstacle
Healthbar
I added a healthbar in the topleft section of the screen

Obstacles
I added a wooden barrier that @ToddBoeker made. The car takes damage when hitting it. I have to still implement the car slowing down when hit by it.

Power-ups
I created the four powersups. When the car hits the powerbox the player receives a random power up if they already have it 20 seconds is added to the buff.

Changes that need to happen to these systems :

- I have to finish the art for all the powerups
- I have to make the car slow down when hit by an obstacle
- I have to make the powerup banners smaller and if I have time make it so when one runs out all the rest shift down the screen